Add repositories and modify DTO

+++ src/main/java/seabattles/repository/UserJpaRepo.java
@@ -2,11 +2,10 @@ package seabattles.repository;
 
 import seabattles.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface UserJpaRepo extends JpaRepository<User, Integer> {
+public interface UserJpaRepo extends JpaRepository<User, Long> {
     User findByLogin(String login);
     User findByLoginAndPassword(String login, String password);
 }
